realigned root folder structure
